fix: [일반유저]내견적관리 - 공통컴포넌트 분리
