+++ src/app/customer/moving-quote/history/features/ReceivedQuote/core/components/ReceivedQuoteCard.tsx
@@ -6,6 +6,7 @@ import { OfferCard } from './OfferCard';
 import { formatToFullDateWithTime, formatToYYMMDD } from '@/shared/utils/dataFormatter';
 import { OfferFilter } from './OfferFilter';
 import { useState } from 'react';
+import { QuoteDetailCard } from '@/app/customer/moving-quote/history/core/components/QuoteDetailCard';
 
 type FilterType = 'all' | 'confirmed';
 
@@ -35,22 +36,7 @@ export const ReceivedQuoteCard = ({ data }: { data: CustomerQuoteHistoryData }) 
 
   return (
     <Stack sx={cardContainerSx}>
-      <Stack sx={dataWrapperSx}>
-        <Typo content="견적 정보" className="text_SB_16to24" color={colorChips.black[400]} />
-        <Stack sx={dataBoxSx}>
-          {quoteData.map((item) => (
-            <Stack key={item.label} direction="row" alignItems="center">
-              <Typo
-                content={item.label}
-                className="text_R_14to20"
-                color={colorChips.grayScale[300]}
-                customStyle={{ width: '110px' }}
-              />
-              <Typo content={item.value} className="text_R_14to20" color={colorChips.black[400]} />
-            </Stack>
-          ))}
-        </Stack>
-      </Stack>
+      <QuoteDetailCard quoteData={quoteData} />
       <Stack sx={dataWrapperSx}>
         <Typo content="견적서 목록" className="text_SB_16to24" color={colorChips.black[400]} />
         <Stack direction="column" width="100%" gap={{ xs: '16px', md: '32px' }}>
@@ -83,17 +69,6 @@ const dataWrapperSx = {
   width: '100%',
   height: '100%',
   gap: { xs: '24px', md: '40px' },
-};
-
-const dataBoxSx = {
-  flexDirection: 'column',
-  width: '100%',
-  height: '100%',
-  gap: '8px',
-  borderRadius: '16px',
-  padding: { xs: '16px', sm: '32px 40px' },
-  backgroundColor: colorChips.background.f7f7f7,
-  border: { xs: 'none', sm: `1px solid ${colorChips.line.f2f2f2}` },
 };
 
 const offersWrapperSx = {
